Unify GetTempFBO
